requirements.txt
pandas
numpy
scikit-learn
streamlit
opencv-python
openai
joblib
Pillow
